EclProblem: move properties to separate header
