+++ CMakeLists_files.cmake
@@ -387,6 +387,7 @@ list (APPEND PUBLIC_HEADER_FILES
   ebos/ecloutputblackoilmodule.hh
   ebos/eclpolyhedralgridvanguard.hh
   ebos/eclproblem.hh
+  ebos/eclproblem_properties.hh
   ebos/eclsolutioncontainers.hh
   ebos/eclthresholdpressure.hh
   ebos/ecltracermodel.hh
